changed float to int in for loop

+++ K&R/chapter1/exercises/1-3.c
@@ -28,9 +28,9 @@ int main(void)
     float celsius = 0;
 
     puts("Fahrenheit -> Celsius");
-    for (float fahr = 0; fahr <= 300; fahr += 20) {
+    for (int fahr = 0; fahr <= 300; fahr += 20) {
         celsius = (5. / 9.) * (fahr - 32.);
-        printf("%3.0f\t\t%5.1f\n", fahr, celsius);
+        printf("%3.0d\t\t%5.1f\n", fahr, celsius);
     }
 
     return 0;
